+++ integrations/zendesk-conversations/src/index.tsx
@@ -6,34 +6,12 @@ import {
 import type { IntegrationInstallation } from '@gitbook/api';
 import { ZendeskRuntimeContext } from './types';
 import { configComponent } from './config';
+import { getZendeskOAuthConfig } from './client';
 
 export default createIntegration<ZendeskRuntimeContext>({
     fetch: (request, context) => {
-        const oauthHandler = createOAuthHandler({
-            redirectURL: `${context.environment.integration.urls.publicEndpoint}/oauth`,
-            clientId: context.environment.secrets.CLIENT_ID,
-            clientSecret: context.environment.secrets.CLIENT_SECRET,
-            scopes: ['tickets:read'],
-            authorizeURL: (installation) => {
-                const subdomain = assertInstallationSubdomain(installation);
-                return `https://${subdomain}.zendesk.com/oauth/authorizations/new`
-            },
-            accessTokenURL: (installation) => {
-                const subdomain = assertInstallationSubdomain(installation);
-                return `https://${subdomain}.zendesk.com/oauth/tokens`
-            },
-        }, { replace: false });
-
+        const oauthHandler = createOAuthHandler(getZendeskOAuthConfig(context), { replace: false });
         return oauthHandler(request, context);
     },
     components: [configComponent]
 });
-
-function assertInstallationSubdomain(installation: IntegrationInstallation) {
-    const subdomain = installation?.configuration?.subdomain;
-    if (!subdomain) {
-        throw new ExposableError('Subdomain is not configured');
-    }
-
-    return subdomain;
-}
